backend/requirements.txt
fastapi
uvicorn
python-dotenv
pandas
sqlalchemy
pymssql
pydantic
python-multipart
psycopg2-binary
